+++ internal/namespaces/k8s/v1beta4/testdata/test-get-version-error.stderr.golden
@@ -1,0 +1,1 @@
+Version test not found
