+++ internal/namespaces/k8s/v1beta4/testdata/test-get-version-simple.stdout.golden
@@ -1,0 +1,22 @@
+region                          fr-par
+name                            1.17.4
+label                           Kubernetes 1.17.4
+available-cnis.0                cilium
+available-cnis.1                calico
+available-cnis.2                weave
+available-cnis.3                flannel
+available-ingresses.0           none
+available-ingresses.1           nginx
+available-ingresses.2           traefik
+available-container-runtimes.0  docker
+available-container-runtimes.1  containerd
+available-container-runtimes.2  crio
+available-feature-gates.0       TTLAfterFinished
+available-feature-gates.1       ServiceTopology
+available-feature-gates.2       EndpointSlice
+available-feature-gates.3       HPAScaleToZero
+available-feature-gates.4       ServiceNodeExclusion
+available-admission-plugins.0   PodSecurityPolicy
+available-admission-plugins.1   MutatingAdmissionWebhook
+available-admission-plugins.2   PodNodeSelector
+available-admission-plugins.3   AlwaysPullImages
